Refactor token response into a hook
